requirements.txt
argparse==1.2.1
distribute==0.6.45
httplib2==0.8
oauth2==1.5.211
python-twitter==1.0
simplejson==3.3.0
ttapi==1.2.0
websocket-client==0.10.0
wsgiref==0.1.2
